backend/requirements.txt
# FastAPI et serveur web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.2
email-validator==2.2.0
python-slugify==8.0.4
python-decouple==3.8

# Traitement de données (versions compatibles Python 3.12)
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
aiofiles==23.2.1

# Validation et sérialisation
pydantic==2.5.2
pydantic-settings==2.1.0

# Visualisations (versions compatibles)
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Logging et monitoring
loguru==0.7.2

# Développement
black==23.11.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# CORS et middleware
fastapi-cors==0.0.6 